requirements.txt
jpype1
tabula-py
pdf2image
boto3